:seedling: Support clayful delete method
